Move rich handling into console module
