+++ anaconda_conda_tos/console/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (C) 2024 Anaconda, Inc
+# SPDX-License-Identifier: BSD-3-Clause
+"""Render functions for console output."""
+
+from .render import list_tos  # noqa: F401
